// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const NodeType = {
  INITIAL: 'INITIAL',
  MANUAL_TRIGGER: 'MANUAL_TRIGGER',
  HTTP_REQUEST: 'HTTP_REQUEST',
  GOOGLE_FORM_TRIGGER: 'GOOGLE_FORM_TRIGGER',
  STRIPE_TRIGGER: 'STRIPE_TRIGGER',
  GEMINI: 'GEMINI',
  DISCORD: 'DISCORD',
  SLACK: 'SLACK'
} as const

export type NodeType = (typeof NodeType)[keyof typeof NodeType]


export const CredentialType = {
  GEMINI: 'GEMINI'
} as const

export type CredentialType = (typeof CredentialType)[keyof typeof CredentialType]


export const ExecutionStatus = {
  RUNING: 'RUNING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED'
} as const

export type ExecutionStatus = (typeof ExecutionStatus)[keyof typeof ExecutionStatus]